requirements.txt
Django>=4.2
Djoser
djangorestframework
python-dotenv
djangorestframework-simplejwt